+++ src/main/java/me/villagerunknown/platform/cmd/flushcachesCommand.java
@@ -1,17 +1,17 @@
 package me.villagerunknown.platform.cmd;
 
-import me.villagerunknown.platform.Platform;
 import me.villagerunknown.platform.feature.commandsFeature;
+import me.villagerunknown.platform.feature.playerCacheFeature;
 import me.villagerunknown.platform.util.MessageUtil;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.text.Text;
 
-public class reloadCommand {
+public class flushcachesCommand {
 	
 	public static void execute() {
-		Platform.reload();
+		playerCacheFeature.emptyCache();
 		
-		String message = MessageUtil.formComment( "Platform is reloading config data..." );
+		String message = MessageUtil.formComment( "Platform flushed caches..." );
 		commandsFeature.COMMAND_SOURCE.sendMessage( Text.of( message ) );
 		commandsFeature.playSound(SoundEvents.ENTITY_VILLAGER_YES);
 	}